# public/api-docs.yaml
openapi: 3.0.0
info:
  title: ScamGuard API
  description: Comprehensive scam detection and prevention API
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/scan:
    post:
      tags: [AI]
      summary: Scan content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: {type: string}
      responses:
        200:
          description: OK
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT